Capstone-ml/django_rest/requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
cached-property==1.5.2
Django==3.2.3
djangorestframework==3.12.4
gast==0.4.0
google-pasta==0.2.0
grpcio==1.38.0
gunicorn==20.1.0
h5py==3.2.1
importlib-metadata==4.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mysql-client==0.0.1
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
protobuf==3.17.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
